filtros, tipos e ajustes

+++ lib/screens/created_pokemons/my_poke_widgets/my_poke_card.dart
@@ -2,7 +2,11 @@ import "package:cloud_firestore/cloud_firestore.dart";
 import "package:flutter/material.dart";
 import 'package:firebase_auth/firebase_auth.dart';
 
-Widget mypokeCard(Function()? onTap, QueryDocumentSnapshot doc) {
+// Defina uma função de retorno de chamada onDelete
+typedef void OnDelete();
+
+Widget mypokeCard(
+    Function()? onTap, QueryDocumentSnapshot doc, OnDelete onDelete) {
   User? user = FirebaseAuth.instance.currentUser;
   return Card(
     child: ListTile(
@@ -12,13 +16,16 @@ Widget mypokeCard(Function()? onTap, QueryDocumentSnapshot doc) {
       onTap: onTap,
       trailing: IconButton(
         icon: Icon(Icons.delete),
-        onPressed: () {
-          final docPoke = FirebaseFirestore.instance
-              .collection('users')
-              .doc(user!.uid)
-              .collection('mypokes')
-              .doc(doc.id);
-          docPoke.delete();
+        onPressed: () async {
+          if (user != null) {
+            final docPoke = FirebaseFirestore.instance
+                .collection('users')
+                .doc(user!.uid)
+                .collection('mypokes')
+                .doc(doc.id);
+            await docPoke.delete();
+            onDelete(); // Chame a função de retorno de chamada após a exclusão
+          }
         },
       ),
     ),
